deplacement gauche droite
