+++ Assets/Scenes/Scene 1.unity
@@ -8128,6 +8128,7 @@ GameObject:
   - component: {fileID: 2029551907}
   - component: {fileID: 2029551906}
   - component: {fileID: 2029551905}
+  - component: {fileID: 2029551908}
   m_Layer: 0
   m_Name: player
   m_TagString: Untagged
@@ -8174,7 +8175,7 @@ SpriteRenderer:
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
   m_SortingLayer: 0
-  m_SortingOrder: 0
+  m_SortingOrder: 3
   m_Sprite: {fileID: -1766820586291129348, guid: 437a04fe29f434a4db1dfb2c6747eb15, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
@@ -8194,7 +8195,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2029551902}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -9.392, y: -0.667, z: 0}
+  m_LocalPosition: {x: -9.392, y: -0.671, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
@@ -8213,7 +8214,7 @@ BoxCollider2D:
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: -0.0040926933, y: 0}
+  m_Offset: {x: -0.0040926933, y: 0.015271902}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
@@ -8224,7 +8225,7 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 0.62186337, y: 0.66}
+  m_Size: {x: 0.62186337, y: 0.62945616}
   m_EdgeRadius: 0
 --- !u!50 &2029551906
 Rigidbody2D:
@@ -8259,3 +8260,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bc51388ea2eb0db4e93fa238572dddd5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  moveSpeed: 200
+  rb: {fileID: 2029551906}
+--- !u!58 &2029551908
+CircleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2029551902}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: -0.03134796}
+  serializedVersion: 2
+  m_Radius: 0.29865205
